[컴파일 불가] 코드 정리
